pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wav2hyp"
version = "1.0.0"
description = "Waveform to Hypocenter processing pipeline for seismic data"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "WAV2HYP Development Team", email = "support@example.com"}
]
maintainers = [
    {name = "WAV2HYP Development Team", email = "support@example.com"}
]
keywords = [
    "seismology", 
    "earthquake", 
    "phase-picking", 
    "event-association", 
    "earthquake-location",
    "machine-learning",
    "volcano-monitoring"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Information Analysis"
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0", 
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "obspy>=1.3.0",
    "seisbench>=0.4.0",
    "pyocto>=0.1.0",
    "torch>=1.9.0",
    "PyYAML>=6.0",
    "h5py>=3.3.0",
    "pyasdf>=0.8.0",
#    "nllpy>=0.1.0",
#    "vdapseisutils>=1.0.0",
    "tqdm>=4.62.0",
    "click>=8.0.0"
]

[project.optional-dependencies]
mapping = ["cartopy>=0.20.0"]
jupyter = ["jupyter>=1.0.0", "ipython>=7.0.0"]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0", 
    "black>=21.0.0",
    "flake8>=3.9.0"
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0"
]
all = [
    "wav2hyp[mapping,jupyter,dev,docs]"
]

[project.urls]
Homepage = "https://github.com/your-org/wav2hyp"
Documentation = "https://wav2hyp.readthedocs.io/"
Repository = "https://github.com/your-org/wav2hyp.git"
Issues = "https://github.com/your-org/wav2hyp/issues"

[project.scripts]
wav2hyp = "wav2hyp.cli:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["wav2hyp*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
"wav2hyp" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--cov=wav2hyp",
    "--cov-report=html",
    "--cov-report=term-missing"
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".tox"
]
